Extract setupMatrixRoom helper
