+++ packages/realm-server/tests/card-endpoints-test.ts
@@ -19,8 +19,7 @@ import {
   setupBaseRealmServer,
   runTestRealmServer,
   setupDB,
-  realmServerTestMatrix,
-  realmSecretSeed,
+  setupMatrixRoom,
   createVirtualNetwork,
   createVirtualNetworkAndLoader,
   matrixURL,
@@ -30,7 +29,6 @@ import {
   waitUntil,
 } from './helpers';
 import '@cardstack/runtime-common/helpers/code-equality-assertion';
-import { MatrixClient } from '@cardstack/runtime-common/matrix-client';
 import { resetCatalogRealms } from '../handlers/handle-fetch-catalog-realms';
 import { APP_BOXEL_REALM_EVENT_TYPE } from '@cardstack/runtime-common/matrix-constants';
 import type {
@@ -69,9 +67,18 @@ module(basename(__filename), function () {
     let request: SuperTest<Test>;
     let dir: DirResult;
 
+    function setTestRequest(newRequest: SuperTest<Test>) {
+      request = newRequest;
+    }
+
+    function getTestRequest() {
+      return request;
+    }
+
     function setupPermissionedRealm(
       hooks: NestedHooks,
       permissions: RealmPermissions,
+      setTestRequest: (newRequest: SuperTest<Test>) => void,
       fileSystem?: Record<string, string | LooseSingleCardDocument>,
     ) {
       setupDB(hooks, {
@@ -99,70 +106,14 @@ module(basename(__filename), function () {
             fileSystem,
           }));
 
-          request = supertest(testRealmHttpServer);
+          setTestRequest(supertest(testRealmHttpServer));
         },
       });
     }
 
-    function setupMatrixRoom(hooks: NestedHooks) {
-      let matrixClient = new MatrixClient({
-        matrixURL: realmServerTestMatrix.url,
-        // it's a little awkward that we are hijacking a realm user to pretend to
-        // act like a normal user, but that's what's happening here
-        username: 'node-test_realm',
-        seed: realmSecretSeed,
-      });
-
-      let testAuthRoomId: string | undefined;
-
-      hooks.beforeEach(async function () {
-        await matrixClient.login();
-        let userId = matrixClient.getUserId()!;
-
-        let response = await request
-          .post('/_server-session')
-          .send(JSON.stringify({ user: userId }))
-          .set('Accept', 'application/json')
-          .set('Content-Type', 'application/json');
-
-        let json = response.body;
-
-        let { joined_rooms: rooms } = await matrixClient.getJoinedRooms();
-
-        if (!rooms.includes(json.room)) {
-          await matrixClient.joinRoom(json.room);
-        }
-
-        await matrixClient.sendEvent(json.room, 'm.room.message', {
-          body: `auth-response: ${json.challenge}`,
-          msgtype: 'm.text',
-        });
-
-        response = await request
-          .post('/_server-session')
-          .send(JSON.stringify({ user: userId, challenge: json.challenge }))
-          .set('Accept', 'application/json')
-          .set('Content-Type', 'application/json');
-
-        testAuthRoomId = json.room;
-
-        await matrixClient.setAccountData('boxel.session-rooms', {
-          [userId]: json.room,
-        });
-      });
-
-      return {
-        matrixClient,
-        getMessagesSince: async function (since: number) {
-          let allMessages = await matrixClient.roomMessages(testAuthRoomId!);
-          let messagesAfterSentinel = allMessages.filter(
-            (m) => m.origin_server_ts > since,
-          );
-
-          return messagesAfterSentinel;
-        },
-      };
-    }
+    hooks.beforeEach(function () {
+      request = getTestRequest();
+    });
 
     let { virtualNetwork, loader } = createVirtualNetworkAndLoader();
 
@@ -185,9 +136,13 @@ module(basename(__filename), function () {
 
     module('card GET request', function (_hooks) {
       module('public readable realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          '*': ['read'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            '*': ['read'],
+          },
+          setTestRequest,
+        );
 
         test('serves the request', async function (assert) {
           let response = await request
@@ -276,9 +231,13 @@ module(basename(__filename), function () {
 
       // using public writable realm to make it easy for test setup for the error tests
       module('public writable realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          '*': ['read', 'write'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            '*': ['read', 'write'],
+          },
+          setTestRequest,
+        );
 
         test('serves a card error request with last known good state', async function (assert) {
           await request
@@ -344,9 +303,13 @@ module(basename(__filename), function () {
       });
 
       module('permissioned realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          john: ['read'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            john: ['read'],
+          },
+          setTestRequest,
+        );
 
         test('401 with invalid JWT', async function (assert) {
           let response = await request
@@ -410,11 +373,15 @@ module(basename(__filename), function () {
 
     module('card POST request', function (_hooks) {
       module('public writable realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          '*': ['read', 'write'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            '*': ['read', 'write'],
+          },
+          setTestRequest,
+        );
 
-        let { getMessagesSince } = setupMatrixRoom(hooks);
+        let { getMessagesSince } = setupMatrixRoom(hooks, getTestRequest);
 
         test('serves the request', async function (assert) {
           let id: string | undefined;
@@ -515,9 +482,13 @@ module(basename(__filename), function () {
       });
 
       module('permissioned realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          john: ['read', 'write'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            john: ['read', 'write'],
+          },
+          setTestRequest,
+        );
 
         test('401 with invalid JWT', async function (assert) {
           let response = await request
@@ -589,11 +560,15 @@ module(basename(__filename), function () {
 
     module('card PATCH request', function (_hooks) {
       module('public writable realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          '*': ['read', 'write'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            '*': ['read', 'write'],
+          },
+          setTestRequest,
+        );
 
-        let { getMessagesSince } = setupMatrixRoom(hooks);
+        let { getMessagesSince } = setupMatrixRoom(hooks, getTestRequest);
 
         test('serves the request', async function (assert) {
           let entry = 'person-1.json';
@@ -743,9 +718,13 @@ module(basename(__filename), function () {
       });
 
       module('permissioned realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          john: ['read', 'write'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            john: ['read', 'write'],
+          },
+          setTestRequest,
+        );
 
         test('401 with invalid JWT', async function (assert) {
           let response = await request
@@ -810,11 +789,15 @@ module(basename(__filename), function () {
 
     module('card DELETE request', function (_hooks) {
       module('public writable realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          '*': ['read', 'write'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            '*': ['read', 'write'],
+          },
+          setTestRequest,
+        );
 
-        let { getMessagesSince } = setupMatrixRoom(hooks);
+        let { getMessagesSince } = setupMatrixRoom(hooks, getTestRequest);
 
         test('serves the request', async function (assert) {
           let entry = 'person-1.json';
@@ -900,9 +883,13 @@ module(basename(__filename), function () {
       });
 
       module('permissioned realm', function (hooks) {
-        setupPermissionedRealm(hooks, {
-          john: ['read', 'write'],
-        });
+        setupPermissionedRealm(
+          hooks,
+          {
+            john: ['read', 'write'],
+          },
+          setTestRequest,
+        );
 
         test('401 with invalid JWT', async function (assert) {
           let response = await request
